requirements.txt
# Core frameworks for agentic logic
langchain==0.1.0
langgraph==0.0.26
langchain-google-genai==1.0.1
google-generativeai==0.3.2

# Web server for API endpoints
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Calendar API
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Email
sendgrid==6.11.0

# Database for persistence
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
